+++ examples/mouse-test.rs
@@ -23,7 +23,9 @@
 extern crate ascii;
 extern crate ncurseswwin;
 
+use std::process::exit;
 use ascii::AsciiChar;
+use anyhow::Result;
 use ncurseswwin::*;
 
 macro_rules! result { ($type: ty) => { Result<$type, NCurseswWinError> } }
@@ -40,15 +42,24 @@ fn main() {
 
         mouse_test(stdscr)
     }) {
-        match source {
-            NCurseswWinError::Panic { message } => eprintln!("panic: {}", message),
-            _                                   => eprintln!("error: {}", source)
+        if let Some(err) = source.downcast_ref::<NCurseswWinError>() {
+            match err {
+                NCurseswWinError::Panic { message } => eprintln!("panic: {}", message),
+                _                                   => eprintln!("error: {}", err)
+            }
+        } else {
+            eprintln!("error: {}", source);
         }
+
+        source.chain().skip(1).for_each(|cause| eprintln!("cause: {}", cause));
+
+        exit(1);
     }
 
+    exit(0);
 }
 
-fn mouse_test(stdscr: &Window) -> result!(()) {
+fn mouse_test(stdscr: &Window) -> Result<()> {
     stdscr.keypad(true)?;
 
     let mut origin = Origin { y: 1, x: 1 };