+++ examples/scr_dump-test.rs
@@ -1,5 +1,5 @@
 /*
-    examples/screen-test.rs
+    examples/scr_dump-test.rs
 
     Copyright (c) 2020 Stephen Whittle  All rights reserved.
 
@@ -22,7 +22,7 @@
 
 extern crate ncurseswwin;
 
-use std::{convert::TryFrom, env, io, process::exit};
+use std::{convert::TryFrom, process::exit, path::Path};
 use anyhow::{Result, Context};
 use ncurseswwin::*;
 
@@ -46,28 +46,26 @@ fn main() {
 }
 
 fn main_routine() -> Result<()> {
-    // We wrap our use of ncurseswin with this function.
-    safe_entry(|| {
-        let term = &env::var("TERM").with_context(|| "$TERM is invalid!!!")?;
+    // We wrap all our use of ncurseswin with this function.
+    ncursesw_entry(|_| {
+        set_input_mode(InputMode::Character)?;
+        set_echo(false)?;
+        set_newline(false)?;
+        intrflush(false)?;
 
-        // create a screen using stdout and stdin for output and input.
-        let screen = &Screen::new(Some(term), io::stdout(), io::stdin())?;
+        cursor_set(CursorType::Invisible)?;
 
-        screen.set_input_mode(InputMode::Character)?;
-        screen.set_echo(false)?;
-        screen.set_newline(false)?;
-        screen.intrflush(false)?;
-
-        // make the screens cursor invisible.
-        screen.cursor_set(CursorType::Invisible)?;
-
-        screen_test(screen)
+        // call our test routine.
+        window_test()
     })
 }
 
-fn screen_test(screen: &Screen) -> Result<()> {
-    // create a window on our screen.
-    let window = &Window::new_sp(screen, Size::default(), Origin::default())?;
+fn window_test() -> Result<()> {
+    // create a window (size default and origin default provided us with
+    // the maximum size of the terminal/display).
+    let window = &Window::new(Size::default(), Origin::default())?;
+
+    window.keypad(true)?;
 
     // extract the box drawing characters for the box drawing type.
     let left_side   = chtype_box_graphic(BoxDrawingGraphic::LeftVerticalLine);
@@ -79,35 +77,65 @@ fn screen_test(screen: &Screen) -> Result<()> {
     let lower_left  = chtype_box_graphic(BoxDrawingGraphic::LowerLeftCorner);
     let lower_right = chtype_box_graphic(BoxDrawingGraphic::LowerRightCorner);
 
-    // create a border on the inital window.
+    // create a border on the window.
     window.border(left_side, right_side, top_side, bottom_side, upper_left, upper_right, lower_left, lower_right)?;
 
     // the text we are going to output.
     let line1 = "If the doors of perception were cleansed every thing would appear to man as it is: Infinite.";
     let line2 = "For man has closed himself up, till he sees all things thro' narrow chinks of his cavern.";
-    let line3 = "Press any key to exit";
+    let line3 = "Press any key to continue to take a screen dump ";
+    let line4 = "A screen dump of the previous screen has been created";
+    let line5 = "Press any key to restore the screen dump";
+    let line6 = "This screen has been restored using scr_restore()";
+    let line7 = "from the first screen displayed by this program.";
 
-    // get the window's size.
     let window_size = window.size()?;
 
     // calculate the initial origin for line 1.
-    let mut origin = Origin { y: (window_size.lines / 2) - 2, x: calc_x_axis(line1, window_size)? };
+    let mut origin = Origin { y: (window_size.lines / 2) - 2, x: calc_x_axis(line1, window_size.columns)? };
 
     // output our lines centered on the x-axis.
     window.mvaddstr(origin, line1)?;
     origin.y += 1;
-    origin.x = calc_x_axis(line2, window_size)?;
+    origin.x = calc_x_axis(line2, window_size.columns)?;
     window.mvaddstr(origin, line2)?;
     origin.y += 2;
-    origin.x = calc_x_axis(line3, window_size)?;
+    origin.x = calc_x_axis(line3, window_size.columns)?;
     window.mvaddstr(origin, line3)?;
 
     // wait for the user to press a key.
     window.getch()?;
 
+    let filename = &Path::new("/tmp/scr_dump.dat");
+
+    scr_dump(filename).with_context(|| errno::errno())?;
+
+    window.clear()?;
+
+    origin = Origin { y: (window_size.lines / 2) - 2, x: calc_x_axis(line4, window_size.columns)? };
+    window.mvaddstr(origin, line4)?;
+    origin.y += 2;
+    origin.x = calc_x_axis(line5, window_size.columns)?;
+    window.mvaddstr(origin, line5)?;
+
+    window.getch()?;
+
+    scr_restore(filename).with_context(|| errno::errno())?;
+
+    doupdate()?;
+
+    origin = Origin { y: window_size.lines - 4, x: calc_x_axis(line6, window_size.columns)? };
+    window.mvaddstr(origin, line6)?;
+    origin.y += 1;
+    origin.x = calc_x_axis(line7, window_size.columns)?;
+    window.mvaddstr(origin, line7)?;
+
+    window.getch()?;
+
     Ok(())
 }
 
-fn calc_x_axis(line: &str, window_size: Size) -> Result<u16> {
-    Ok((window_size.columns / 2) - (u16::try_from(line.len())? / 2))
+// calculate a centered x-axis based on the length of the string we are outputing.
+fn calc_x_axis(line: &str, columns: u16) -> Result<u16> {
+    Ok((columns / 2) - (u16::try_from(line.len())? / 2))
 }