+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All breaking changes are marked with [BC] and potentially require API consumer c
 ## [0.6.0] - ????-??-?? [BC]
 - NCurses form module implemented as `form`.
 - Now uses `thiserror` crate instead of `custom_error` crate to create `Error` types.
+- `ncursesw_entry()`, `ncursesw_init()` and `safe_entry()` now use `anyhow::Result` as a return type. [BC]
 - Added NCurses screen functionality.
 - Added `{normal,extend}::SoftLabels`.
 - Removed `setlocale()` function. [BC]